pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "mpvremote"
version = "0.0.1"
dependencies = [
    "pyserial",
    "pyautogui", # only used in the example config.py
]

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[project.scripts]
mpvremote = "mpvremote.cli:main"
